introduced #has_widgets which allows adding widgets in controller class context. widgets that should be existing in multiple actions can thus be defined directly in controller classes, like
class HunterController < ApplicationController::Base
  has_widgets do |root|
    root << cell(:bear_trap, :charged, 'nasty_trap')
  end

  def trap_bears
    @box = render_widget 'nasty_trap'

removed the controller reference from the StatefulWidget constructor. that's a major API change. the controller is instead passed before #invoke_state, and it's planned to pass it as a real #invoke_state argument. widgets don't need a controller except for rendering.
removed the error-prone ApplicationWidgetTree, it's no longer processed. to define an initial app tree structure, use #has_widgets in the ApplicationController.